requirements.txt
torch
pillow
torchvision
wandb
matplotlib